+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Change
 * let-box syntax `let box[Ψ.U] = e1 in e2` -> `let box[U] = e1 in e2`
 * clo syntax `U with (e1, ..., en)` -> `U with [e1/x1, ..., en/xn]`
 
+Fix
+* global substitution
+
 ## 0.2
 Add
 * bool, integer, product, and list type